Prevent circular dependencies
